Move Streams/Prelude.hs to Internal/Data/Stream
