+++ src/Streamly/Internal/Data/Pipe.hs
@@ -269,7 +269,7 @@ import Streamly.Internal.Data.Pipe.Types
 -- import qualified Streamly.Prelude as S
 -- import qualified Streamly.Internal.Data.Stream.StreamD as D
 -- import qualified Streamly.Internal.Data.Stream.StreamK as K
--- import qualified Streamly.Streams.Prelude as P
+-- import qualified Streamly.Internal.Data.Stream.Prelude as P
 
 ------------------------------------------------------------------------------
 -- Pipes
